refactor: fold renderer_settings, template_load, render_context into core/renderer.py
Renderer absorbs its process-wide settings (no more RendererSettings base
class), template resolution, render-context building, and reactive stamping.
Behavior unchanged.
